Fix inability to find plain.mem when filename is passed on the command line
An off-by-one error in mp_pack_buffered_name() usage caused a command such as
'./newmpost test.mp' to try to dump 'plaint.mem' instead of 'plain.mem' because
the first character of the filename passed on the command line was appended to
the format file name.

Fix by using a more C-style call convention for mp_pack_buffered_name().


git-svn-id: https://serveur-svn.lri.fr/svn/modhel/metapost@301 b8f95e4b-c8ec-496e-911b-43368900ac28